+++ src/ui.rs
@@ -26,6 +26,14 @@ pub fn render(frame: &mut Frame, model: &Model) {
     }
 }
 
+fn connection_status(model: &Model) -> Paragraph {
+    Paragraph::new("● ").style(Style::new().fg(if model.connected {
+        Color::Green
+    } else {
+        Color::Red
+    }))
+}
+
 fn render_topics(frame: &mut Frame, area: Rect, model: &Model, filter: Option<&Filter>) {
     let [top, overview, prompt] = Layout::vertical([
         Length(1),
@@ -34,8 +42,18 @@ fn render_topics(frame: &mut Frame, area: Rect, model: &Model, filter: Option<&F
     ])
     .areas(area);
 
+    let [counter, host, indicator] = Layout::horizontal([Length(6), Fill(0), Length(6)]).areas(top);
+
     // Top header
-    frame.render_widget(Paragraph::new(format!("Messages: {}", model.counter)), top);
+    frame.render_widget(
+        Paragraph::new(format!("{}", model.counter)).dark_gray(),
+        counter,
+    );
+    frame.render_widget(
+        Paragraph::new(model.broker().to_string()).centered().bold(),
+        host,
+    );
+    frame.render_widget(connection_status(model).right_aligned(), indicator);
 
     // Topic overview
     let list = List::new(model.topics().map(|(topic, message)| {
@@ -53,7 +71,7 @@ fn render_topics(frame: &mut Frame, area: Rect, model: &Model, filter: Option<&F
     }))
     .block(
         Block::new()
-            .title(Line::raw("Topics").centered())
+            .title(Line::raw("Topics").italic().gray())
             .borders(Borders::TOP),
     );
 
@@ -83,23 +101,19 @@ fn render_details(
     scroll: u16,
 ) {
     let [header, pane, footer] = Layout::vertical([
-        Length(2),
+        Length(1),
         Fill(0),
         Length(if jq.is_dormant() { 0 } else { 6 }),
     ])
     .areas(area);
+
+    let [header, indicator] = Layout::horizontal([Fill(0), Length(2)]).areas(header);
+
     let [details, scroller] = Layout::horizontal([Fill(0), Length(1)]).areas(pane);
 
     // Top header with topic name
-    frame.render_widget(
-        Paragraph::new(
-            Span::styled(topic, Style::new().fg(Color::Gray))
-                .italic()
-                .into_centered_line(),
-        )
-        .block(Block::new().borders(Borders::BOTTOM)),
-        header,
-    );
+    frame.render_widget(Paragraph::new(topic).bold().centered(), header);
+    frame.render_widget(connection_status(model).right_aligned(), indicator);
 
     let mut style = Style::new();
     if model.is_copy() {
@@ -108,7 +122,13 @@ fn render_details(
 
     let message = model.message(topic).unwrap_or_default();
     frame.render_widget(
-        Paragraph::new(model.highlight(&message, details, scroll).style(style)).scroll((scroll, 0)),
+        Paragraph::new(model.highlight(&message, details, scroll).style(style))
+            .scroll((scroll, 0))
+            .block(
+                Block::new()
+                    .title(Line::raw("Message").italic().gray())
+                    .borders(Borders::TOP),
+            ),
         details,
     );
     frame.render_stateful_widget(