+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# Unreleased
+# CHANGELOG
+
+## Unreleased
+
+* Switch to [rumqttc](https://docs.rs/rumqttc/latest/rumqttc/)
+* Show connection status in header
+* Automatically reconnect if broker restarts
+
+## 1.0.0
 
 * Show topics in selectable list
 * Fuzzy search through topics